Feature/correcao dos bugs

+++ src/main/java/br/edu/utfpr/autorepairshop/model/Maintenance.java
@@ -1,5 +1,6 @@
 package br.edu.utfpr.autorepairshop.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -32,9 +33,9 @@ public class Maintenance {
 	@ManyToOne
 	private Vehicle vehicle;
 	@Column(name = "total_price")
-	private Long totalPrice;
+	private BigDecimal totalPrice;
 	@Column(name = "price_hand_work")
-	private Long priceHandWork;
+	private BigDecimal priceHandWork;
 	private Date date;
 	private String hour;
 	private String description;
@@ -43,10 +44,10 @@ public class Maintenance {
 	// no banco será
 	// implementado a modelagem das peças para uma futura atualização.
 	@Column(name = "price_parts")
-	private Long priceParts;
+	private BigDecimal priceParts;
 
 	public Maintenance(Client client, Employee employee, AutoRepairShop autoRepairShop,
-			Vehicle vehicle, Long totalPrice, Long priceHandWork, Long priceParts, Date date, String hour,
+			Vehicle vehicle, BigDecimal totalPrice, BigDecimal priceHandWork, BigDecimal priceParts, Date date, String hour,
 			String description) {
 		this.client = client;
 		this.employee = employee;
